Add manifest version to manifest file
In order to avoid getting an error from the latest versions of chrome
when loading the unpacked package, a manifest version is required. This
is not specific to this extension, it's specific to the version of the
Chrome API that we're using, so it doesn't need to be updated
frequently, and will be the integer 2 until some point in the future
that we decide to use a new manifest type.