Redirect to the todo list when browsing to twitter.com
When a user opens a new tab, chrome fires the "Tab Updated" event. I've
added a background script that listens for that event, and if the the
new url contains "twitter.com", we update the url of that new tab to our
todo list at "newtab.html."

The extension previously worked by replacing our newtab page with the
todo list using the "chrome_url_overrides" feature, so I replaced that
with the background script, which also required that we ask for the tabs
permission so we can exercise the chrome.tabs interface.

I've also removed the update_url, since these changes would be wiped out
when the original branch bumps it's version.

On consern I have is that we're setting a new listener everytime the
browser is throwing events. I don't think this is the case, but I've
previously used a background page instead of background scripts, so I'm
not as familiar with how these work. Could be totally fine! Truly,
there's no way to know.